+++ discord/cogs/csvmembers/__init__.py
@@ -1,0 +1,5 @@
+from .csvmembers import CsvMembers
+
+
+async def setup(bot):
+    await bot.add_cog(CsvMembers())